Comentarios normalizados

+++ Backend/src/main/java/edu/usach/tbdgrupo5/entities/Artista.java
@@ -20,9 +20,9 @@ public class Artista implements Serializable {
 	@Id
 	private int idartista;
 
-	private int comentariosNegativos;
+	private double comentariosNegativos;
 
-	private int comentariosPositivos;
+	private double comentariosPositivos;
 
 	private int comentariosNeutros;
 
@@ -63,19 +63,19 @@ public class Artista implements Serializable {
 		this.idartista = idartista;
 	}
 
-	public int getComentariosNegativos() {
+	public double getComentariosNegativos() {
 		return this.comentariosNegativos;
 	}
 
-	public void setComentariosNegativos(int comentariosNegativos) {
+	public void setComentariosNegativos(double comentariosNegativos) {
 		this.comentariosNegativos = comentariosNegativos;
 	}
 
-	public int getComentariosPositivos() {
+	public double getComentariosPositivos() {
 		return this.comentariosPositivos;
 	}
 
-	public void setComentariosPositivos(int comentariosPositivos) {
+	public void setComentariosPositivos(double comentariosPositivos) {
 		this.comentariosPositivos = comentariosPositivos;
 	}
